make SuiteClient available only in the test environment (#232)
